[release/2.4] Fix for test_c10d_nccl.py:test_comm_split_subgroup
FBCODE internally uses ncclCommCreateFromRanks in nccl version 2.21+, which doesn't include inactive ranks in the split and therefore the expected count at this point is 0 for all ranks.

However, externally even for nccl version 2.21+, we use ncclCommSplit, which includes all ranks in the split, the inactive ones are used with NOCOLOR. Therefore, rank 0 will have a split of 0 as it has not done a comm yet and rank 1 will have a split of 1 (inactive though).

This change allows the UT to run for both cases.